refactor(gen_reply): move link-context builders into the link_sources package
